Extract classes.
