+++ nextcloud/src/main/java/ch/cyberduck/core/nextcloud/NextcloudShareFeature.java
@@ -24,42 +24,36 @@ import ch.cyberduck.core.LoginOptions;
 import ch.cyberduck.core.PasswordCallback;
 import ch.cyberduck.core.Path;
 import ch.cyberduck.core.PathRelativizer;
-import ch.cyberduck.core.StringAppender;
 import ch.cyberduck.core.URIEncoder;
 import ch.cyberduck.core.dav.DAVSession;
 import ch.cyberduck.core.exception.BackgroundException;
-import ch.cyberduck.core.features.Share;
 import ch.cyberduck.core.http.DefaultHttpResponseExceptionMappingService;
+import ch.cyberduck.core.ocs.OcsDownloadShareResponseHandler;
+import ch.cyberduck.core.ocs.OcsShareeResponseHandler;
+import ch.cyberduck.core.ocs.OcsUploadShareResponseHandler;
+import ch.cyberduck.core.ocs.model.Share;
 
 import org.apache.commons.lang3.StringUtils;
 import org.apache.http.HttpEntity;
 import org.apache.http.HttpHeaders;
-import org.apache.http.HttpResponse;
-import org.apache.http.StatusLine;
 import org.apache.http.client.HttpResponseException;
 import org.apache.http.client.methods.HttpGet;
 import org.apache.http.client.methods.HttpPost;
 import org.apache.http.client.methods.HttpPut;
-import org.apache.http.entity.BufferedHttpEntity;
 import org.apache.http.entity.ContentType;
-import org.apache.http.impl.client.AbstractResponseHandler;
-import org.apache.http.util.EntityUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
 import java.io.IOException;
-import java.net.URI;
 import java.text.MessageFormat;
 import java.util.Collections;
-import java.util.HashSet;
 import java.util.LinkedHashSet;
 import java.util.Set;
 
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.dataformat.xml.XmlMapper;
 import com.github.sardine.impl.handler.VoidResponseHandler;
 
-public class NextcloudShareFeature implements Share {
+public class NextcloudShareFeature implements ch.cyberduck.core.features.Share {
     private static final Logger log = LogManager.getLogger(NextcloudShareFeature.class);
 
     private final DAVSession session;
@@ -107,26 +101,7 @@ public class NextcloudShareFeature implements Share {
                         Collections.singletonList(Sharee.world)
                 );
                 try {
-                    sharees.addAll(session.getClient().execute(resource, new OcsResponseHandler<Set<Sharee>>() {
-                                @Override
-                                public Set<Sharee> handleEntity(final HttpEntity entity) throws IOException {
-                                    final XmlMapper mapper = new XmlMapper();
-                                    final ocs value = mapper.readValue(entity.getContent(), ocs.class);
-                                    if(value.data != null) {
-                                        if(value.data.users != null) {
-                                            final Set<Sharee> sharees = new HashSet<>();
-                                            for(ocs.user user : value.data.users) {
-                                                final String id = user.value.shareWith;
-                                                final String label = String.format("%s (%s)", user.label, user.shareWithDisplayNameUnique);
-                                                sharees.add(new Sharee(id, label));
-                                            }
-                                            return sharees;
-                                        }
-                                    }
-                                    return Collections.emptySet();
-                                }
-                            }
-                    ));
+                    sharees.addAll(session.getClient().execute(resource, new OcsShareeResponseHandler()));
                 }
                 catch(HttpResponseException e) {
                     log.warn(String.format("Failure %s retrieving sharees", e));
@@ -163,19 +138,7 @@ public class NextcloudShareFeature implements Share {
         resource.setHeader("OCS-APIRequest", "true");
         resource.setHeader(HttpHeaders.ACCEPT, ContentType.APPLICATION_XML.getMimeType());
         try {
-            return session.getClient().execute(resource, new OcsResponseHandler<DescriptiveUrl>() {
-                @Override
-                public DescriptiveUrl handleEntity(final HttpEntity entity) throws IOException {
-                    final XmlMapper mapper = new XmlMapper();
-                    final ocs value = mapper.readValue(entity.getContent(), ocs.class);
-                    if(null != value.data) {
-                        if(null != value.data.url) {
-                            return new DescriptiveUrl(URI.create(value.data.url), DescriptiveUrl.Type.http);
-                        }
-                    }
-                    return DescriptiveUrl.EMPTY;
-                }
-            });
+            return session.getClient().execute(resource, new OcsDownloadShareResponseHandler());
         }
         catch(HttpResponseException e) {
             throw new DefaultHttpResponseExceptionMappingService().map(e);
@@ -207,27 +170,22 @@ public class NextcloudShareFeature implements Share {
         resource.setHeader("OCS-APIRequest", "true");
         resource.setHeader(HttpHeaders.ACCEPT, ContentType.APPLICATION_XML.getMimeType());
         try {
-            return session.getClient().execute(resource, new OcsResponseHandler<DescriptiveUrl>() {
+            return session.getClient().execute(resource, new OcsUploadShareResponseHandler() {
                 @Override
                 public DescriptiveUrl handleEntity(final HttpEntity entity) throws IOException {
                     final XmlMapper mapper = new XmlMapper();
-                    final ocs value = mapper.readValue(entity.getContent(), ocs.class);
-                    if(null != value.data) {
-                        // Additional request, because permissions are ignored in POST
-                        final StringBuilder request = new StringBuilder(String.format("https://%s/ocs/v1.php/apps/files_sharing/api/v1/shares/%s?permissions=%d",
-                                bookmark.getHostname(),
-                                value.data.id,
-                                SHARE_PERMISSIONS_CREATE
-                        ));
-                        final HttpPut put = new HttpPut(request.toString());
-                        put.setHeader("OCS-APIRequest", "true");
-                        put.setHeader(HttpHeaders.ACCEPT, ContentType.APPLICATION_XML.getMimeType());
-                        session.getClient().execute(put, new VoidResponseHandler());
-                        if(null != value.data.url) {
-                            return new DescriptiveUrl(URI.create(value.data.url), DescriptiveUrl.Type.http);
-                        }
-                    }
-                    return DescriptiveUrl.EMPTY;
+                    final Share value = mapper.readValue(entity.getContent(), Share.class);
+                    // Additional request, because permissions are ignored in POST
+                    final StringBuilder request = new StringBuilder(String.format("https://%s/ocs/v1.php/apps/files_sharing/api/v1/shares/%s?permissions=%d",
+                            bookmark.getHostname(),
+                            value.data.id,
+                            SHARE_PERMISSIONS_CREATE
+                    ));
+                    final HttpPut put = new HttpPut(request.toString());
+                    put.setHeader("OCS-APIRequest", "true");
+                    put.setHeader(HttpHeaders.ACCEPT, ContentType.APPLICATION_XML.getMimeType());
+                    session.getClient().execute(put, new VoidResponseHandler());
+                    return super.handleEntity(entity);
                 }
             });
         }
@@ -239,112 +197,4 @@ public class NextcloudShareFeature implements Share {
         }
     }
 
-    /*
-    <ocs>
-     <meta>
-      <status>ok</status>
-      <statuscode>200</statuscode>
-      <message>OK</message>
-     </meta>
-     <data>
-      <id>36</id>
-      <share_type>3</share_type>
-      <uid_owner>dkocher</uid_owner>
-      <displayname_owner>David Kocher</displayname_owner>
-      <permissions>1</permissions>
-      <stime>1559218292</stime>
-      <parent/>
-      <expiration/>
-      <token>79NKo6JxmsxxGBb</token>
-      <uid_file_owner>dkocher</uid_file_owner>
-      <note></note>
-      <label></label>
-      <displayname_file_owner>David Kocher</displayname_file_owner>
-      <path>/sandbox/example.png</path>
-      <item_type>file</item_type>
-      <mimetype>image/png</mimetype>
-      <storage_id>home::dkocher</storage_id>
-      <storage>3</storage>
-      <item_source>36285</item_source>
-      <file_source>36285</file_source>
-      <file_parent>36275</file_parent>
-      <file_target>/Monte Panarotta.png</file_target>
-      <share_with/>
-      <share_with_displayname/>
-      <password/>
-      <send_password_by_talk></send_password_by_talk>
-      <url>https://example.net/s/67hgsdfjkds67</url>
-      <mail_send>1</mail_send>
-      <hide_download>0</hide_download>
-     </data>
-    </ocs>
-    */
-    @JsonIgnoreProperties(ignoreUnknown = true)
-    private static final class ocs {
-        public meta meta;
-        public data data;
-
-        @JsonIgnoreProperties(ignoreUnknown = true)
-        private static final class meta {
-            public String status;
-            public String statuscode;
-            public String message;
-            public int itemsperpage;
-            public int totalitems;
-        }
-
-        @JsonIgnoreProperties(ignoreUnknown = true)
-        private static final class data {
-            public String id;
-            public String url;
-            public user[] users;
-        }
-
-        @JsonIgnoreProperties(ignoreUnknown = true)
-        private static final class users {
-            public user[] element;
-        }
-
-        @JsonIgnoreProperties(ignoreUnknown = true)
-        private static final class user {
-            public String label;
-            public String icon;
-            public String shareWithDisplayNameUnique;
-            public value value;
-        }
-
-        @JsonIgnoreProperties(ignoreUnknown = true)
-        private static final class value {
-            public int shareType;
-            public String shareWith;
-            public String shareWithAdditionalInfo;
-        }
-    }
-
-    private abstract static class OcsResponseHandler<R> extends AbstractResponseHandler<R> {
-        @Override
-        public R handleResponse(final HttpResponse response) throws IOException {
-            final StatusLine statusLine = response.getStatusLine();
-            EntityUtils.updateEntity(response, new BufferedHttpEntity(response.getEntity()));
-            if(statusLine.getStatusCode() >= 300) {
-                final StringAppender message = new StringAppender();
-                message.append(statusLine.getReasonPhrase());
-                final ocs error = new XmlMapper().readValue(response.getEntity().getContent(), ocs.class);
-                message.append(error.meta.message);
-                throw new HttpResponseException(statusLine.getStatusCode(), message.toString());
-            }
-            final ocs error = new XmlMapper().readValue(response.getEntity().getContent(), ocs.class);
-            try {
-                if(Integer.parseInt(error.meta.statuscode) > 100) {
-                    final StringAppender message = new StringAppender();
-                    message.append(error.meta.message);
-                    throw new HttpResponseException(Integer.parseInt(error.meta.statuscode), message.toString());
-                }
-            }
-            catch(NumberFormatException e) {
-                log.warn(String.format("Failure parsing status code in response %s", error));
-            }
-            return super.handleResponse(response);
-        }
-    }
 }
